trademarkImage-ml/requirements.txt
awscli==1.11.13
beautifulsoup4==4.4.1
bleach==1.5.0
blinker==1.3
boto3==1.5.28
botocore==1.8.42
chardet==2.3.0
click==6.7
colorama==0.3.7
configobj==5.0.6
cycler==0.10.0
defer==1.0.4
docutils==0.12
enum34==1.1.6
Flask==0.12.2
h5py==2.7.1
html5lib==0.9999999
idna==2.0
itsdangerous==0.24
Jinja2==2.8
jmespath==0.9.0
jsonpatch==1.10
jsonpointer==1.9
Keras==2.3.1
Markdown==2.6.10
MarkupSafe==0.23
matplotlib==2.1.1
numpy==1.14.3
oauthlib==1.0.3
opencv-python==3.3.1.11
pandas==0.22.0
Pillow==8.0.1
prettytable==0.7.2
protobuf==3.5.1
pyasn1==0.1.9
Pygments==2.1
PyJWT==1.3.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.3
pyxdg==0.25
PyYAML==3.11
roman==2.0.0
rsa==3.2.3
s3transfer==0.1.12
scikit-learn==0.19.1
scipy==1.0.0
six==1.11.0
sklearn==0.0
tensorflow==1.13.2
urllib3==1.13.1
virtualenv==15.0.1
Werkzeug==0.13
nltk==3.2.5
flask-cors==3.0.4
gunicorn==19.8.1
pybind11==2.2.3
pymysql==0.10.1
annoy==1.16.0
